requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
bs4==0.0.2
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
chromium==0.0.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django==5.0.6
django-admin-rangefilter==0.13.1
django-celery-beat==2.6.0
django-modeladmin-reorder==0.3.1
django-timezone-field==7.0
exceptiongroup==1.2.2
fake-useragent==1.5.1
greenlet==3.0.3
h11==0.14.0
idna==3.7
kombu==5.3.7
outcome==1.3.0.post0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pyee==11.1.0
PySocks==1.7.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.4
requests==2.32.3
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
undetected-chromedriver==3.5.5
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
whitenoise==6.7.0
wsproto==1.2.0
git+https://github.com/bloodwithmilk25/django-modeladmin-reorder.git@master
